requirements.txt
Django>=4.2,<5.0
gunicorn
whitenoise
dj-database-url
psycopg2-binary
Pillow
torch
torchvision
numpy
os
